+++ tools/benchmark/src/Reporter.ts
@@ -30,8 +30,6 @@ SOFTWARE.
 // This file is a reporter used with node, so depending on node is fine.
 /* eslint-disable import-x/no-nodejs-modules */
 
-/* eslint-disable unicorn/prefer-module */
-
 /* eslint no-console: ["error", { allow: ["log"] }] */
 import * as fs from "node:fs";
 import * as path from "node:path";